feat: Enhance project configuration loading with API credential validation and improve BrowserStackConnect component
